requirements.txt
kafka-python
requests
beautifulsoup4
vaderSentiment
psycopg2-binary
dash
